[console.py, __init__.py, file_storage.py] Refactor to use supported_classes for class validation and instantiation
