+++ models/place.py
@@ -17,7 +17,3 @@ class Place(BaseModel):
     latitude = 0.0
     longitude = 0.0
     amenity_ids: List[str] = []
-
-
-
-5
